need to include the override for shared_ptr in knowrob shared library in case python code is executed from C
